notes on how to change data spread. I think it will be easily to separate nutrient data and field parameters bc nutrient data has reps and field parameters don't.
I initially had data with all parameters in a column but now i've made each parameter a column. May change this but I want to see for now bc there was just a ton of redundant info.

I made two csv files now, one for each, but I want to save this code for reference on how to change spread and save to a csv